Fix graphical artifact in the fisheye effect
The fisheye effect generates a graphical artifact close to the center of
the picture, due to bad precision and division by zero in the shader.
The problem is fixed by making a small change in the shader, so that the
picture is uniformly scaled close to the center instead. This avoids the
problem and looks as expected, without affecting the performance.

Bug: 64107054

Test: Manual - Install dev sample HelloEffects and use the 'fisheye'
from the overflow.  There is a small artifact in the center of the puppy
without this patch applied. Think of the puppies!!
Change-Id: I063f60facd30708db29ff544fdb47ac896e3d54b

Signed-off-by: mydongistiny <jaysonedson@gmail.com>